problem 10 with extended primes size

+++ src/problems/p007.rs
@@ -1,5 +1,5 @@
 use project_euler::primes::primes;
 
 pub fn solve() -> String {
-    primes().skip(10000).take(1).sum::<i32>().to_string()
+    primes().skip(10000).take(1).sum::<u128>().to_string()
 }
